fix(dataDeletion): SAV-558: Pass custom hooks to parent model (#5740)
* sav-558: pass custom hooks from child model to parent model class

* Prevent infinite loop

* Rename variable

* Only add hooks for public schema models

* Ensure hooks are run inside transactions

---------

+++ packages/shared/src/utils/beforeDestroyHooks.js
@@ -7,8 +7,8 @@ async function getIds(options) {
     return ids;
   }
 
-  const encounters = await options.model.findAll(options);
-  return encounters.map(x => x.id);
+  const instances = await options.model.findAll(options);
+  return instances.map(x => x.id);
 }
 
 function getDependantAssociations(model) {
@@ -17,7 +17,17 @@ function getDependantAssociations(model) {
   );
 }
 
-export async function genericBeforeDestroy(instance) {
+async function executeInsideTransaction(sequelize, arg, fn) {
+  if (sequelize.isInsideTransaction()) {
+    await fn(arg);
+    return;
+  }
+  await sequelize.transaction(async () => {
+    await fn(arg);
+  });
+}
+
+async function beforeDestroy(instance) {
   const dependantAssociations = getDependantAssociations(instance.constructor);
 
   for (const association of dependantAssociations) {
@@ -26,12 +36,26 @@ export async function genericBeforeDestroy(instance) {
   }
 }
 
-export async function genericBeforeBulkDestroy(options) {
+async function beforeBulkDestroy(options) {
   const ids = await getIds(options);
+  if (ids.length === 0) {
+    return;
+  }
+
   const dependantAssociations = getDependantAssociations(options.model);
 
   for (const association of dependantAssociations) {
     const { target, foreignKey } = association;
     await target.destroy({ where: { [foreignKey]: { [Op.in]: ids } } });
   }
+}
+
+export async function genericBeforeDestroy(instance) {
+  const { sequelize } = instance;
+  await executeInsideTransaction(sequelize, instance, beforeDestroy);
+}
+
+export async function genericBeforeBulkDestroy(options) {
+  const { sequelize } = options.model;
+  await executeInsideTransaction(sequelize, options, beforeBulkDestroy);
 }
